修复posting字段类型问题：warehouse_id和delivery_method_id改为BigInt
- 将OzonPosting模型的warehouse_id和delivery_method_id从Integer改为BigInteger
- OZON API返回的ID可能超出int32范围（如1020002714275000）
- 创建数据库迁移更新字段类型
- 解决订单同步时的"value out of int32 range"错误

+++ plugins/ef/channels/ozon/models/orders.py
@@ -169,11 +169,11 @@ class OzonPosting(Base):
     
     # 发货信息
     shipment_date = Column(DateTime(timezone=True))
-    delivery_method_id = Column(Integer)
+    delivery_method_id = Column(BigInteger)  # OZON API可能返回超大ID
     delivery_method_name = Column(String(200))
-    
+
     # 仓库
-    warehouse_id = Column(Integer)
+    warehouse_id = Column(BigInteger)  # OZON API可能返回超大ID
     warehouse_name = Column(String(200))
     
     # 包裹信息
